chore: move calendar and event related types to their own file to clean up common.ts per ticket 262
